requirements.txt
allure-pytest==2.9.44
allure-python-commons==2.8.40
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
et-xmlfile==1.1.0
idna==3.2
iniconfig==1.1.1
openpyxl==3.0.7
packaging==21.0
pdfkit==0.6.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-html==3.1.1
pytest-metadata==1.11.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
webdriver-manager==3.5.0
requests
ddt~=1.4.4
pandas~=1.4.1
loguru~=0.6.0
python-dotenv~=0.19.2
randomuser~=1.6